// src/lib/interfaces/rustplus.ts
export interface AppTeamInfo {
    /**
     * @generated from protobuf field: uint64 leaderSteamId = 1
     */
    leaderSteamId: string;
    /**
     * @generated from protobuf field: repeated rustplus.AppTeamInfo.Member members = 2
     */
    members: AppTeamInfo_Member[];
    /**
     * @generated from protobuf field: repeated rustplus.AppTeamInfo.Note mapNotes = 3
     */
    mapNotes: AppTeamInfo_Note[];
    /**
     * @generated from protobuf field: repeated rustplus.AppTeamInfo.Note leaderMapNotes = 4
     */
    leaderMapNotes: AppTeamInfo_Note[];
}
/**
 * @generated from protobuf message rustplus.AppTeamInfo.Member
 */
export interface AppTeamInfo_Member {
    /**
     * @generated from protobuf field: uint64 steamId = 1
     */
    steamId: string;
    /**
     * @generated from protobuf field: string name = 2
     */
    name: string;
    /**
     * @generated from protobuf field: float x = 3
     */
    x: number;
    /**
     * @generated from protobuf field: float y = 4
     */
    y: number;
    /**
     * @generated from protobuf field: bool isOnline = 5
     */
    isOnline: boolean;
    /**
     * @generated from protobuf field: uint32 spawnTime = 6
     */
    spawnTime: number;
    /**
     * @generated from protobuf field: bool isAlive = 7
     */
    isAlive: boolean;
    /**
     * @generated from protobuf field: uint32 deathTime = 8
     */
    deathTime: number;
}
/**
 * @generated from protobuf message rustplus.AppTeamInfo.Note
 */
export interface AppTeamInfo_Note {
    /**
     * @generated from protobuf field: rustplus.AppTeamInfo.Note.Type type = 2
     */
    type: AppTeamInfo_Note_Type;
    /**
     * @generated from protobuf field: float x = 3
     */
    x: number;
    /**
     * @generated from protobuf field: float y = 4
     */
    y: number;
    /**
     * @generated from protobuf field: rustplus.AppTeamInfo.Note.Icon icon = 5
     */
    icon: AppTeamInfo_Note_Icon;
    /**
     * @generated from protobuf field: rustplus.AppTeamInfo.Note.ColourIndex colourIndex = 6
     */
    colourIndex: AppTeamInfo_Note_ColourIndex;
    /**
     * @generated from protobuf field: optional string label = 7
     */
    label?: string;
}
/**
 * @generated from protobuf enum rustplus.AppTeamInfo.Note.Type
 */
export enum AppTeamInfo_Note_Type {
    /**
     * @generated from protobuf enum value: Death = 0;
     */
    Death = 0,
    /**
     * @generated from protobuf enum value: PointOfInterest = 1;
     */
    PointOfInterest = 1
}
/**
 * @generated from protobuf enum rustplus.AppTeamInfo.Note.Icon
 */
export enum AppTeamInfo_Note_Icon {
    /**
     * @generated from protobuf enum value: Pin = 0;
     */
    Pin = 0,
    /**
     * @generated from protobuf enum value: Dollar = 1;
     */
    Dollar = 1,
    /**
     * @generated from protobuf enum value: Home = 2;
     */
    Home = 2,
    /**
     * @generated from protobuf enum value: Parachute = 3;
     */
    Parachute = 3,
    /**
     * @generated from protobuf enum value: Scope = 4;
     */
    Scope = 4,
    /**
     * @generated from protobuf enum value: Shield = 5;
     */
    Shield = 5,
    /**
     * @generated from protobuf enum value: Skull = 6;
     */
    Skull = 6,
    /**
     * @generated from protobuf enum value: Sleep = 7;
     */
    Sleep = 7,
    /**
     * @generated from protobuf enum value: Zzz = 8;
     */
    Zzz = 8,
    /**
     * @generated from protobuf enum value: Gun = 9;
     */
    Gun = 9,
    /**
     * @generated from protobuf enum value: Rock = 10;
     */
    Rock = 10,
    /**
     * @generated from protobuf enum value: Loot = 11;
     */
    Loot = 11
}
/**
 * @generated from protobuf enum rustplus.AppTeamInfo.Note.ColourIndex
 */
export enum AppTeamInfo_Note_ColourIndex {
    /**
     * @generated from protobuf enum value: Yellow = 0;
     */
    Yellow = 0,
    /**
     * @generated from protobuf enum value: Blue = 1;
     */
    Blue = 1,
    /**
     * @generated from protobuf enum value: Green = 2;
     */
    Green = 2,
    /**
     * @generated from protobuf enum value: Red = 3;
     */
    Red = 3,
    /**
     * @generated from protobuf enum value: Purple = 4;
     */
    Purple = 4,
    /**
     * @generated from protobuf enum value: Cyan = 5;
     */
    Cyan = 5
}
// ...
export const AppTeamInfo = new AppTeamInfo$Type();
// ...
export const AppTeamInfo_Member = new AppTeamInfo_Member$Type();
// @generated message type with reflection information, may provide speed optimized methods
class AppTeamInfo_Note$Type extends MessageType<AppTeamInfo_Note> {
    constructor() {
        super("rustplus.AppTeamInfo.Note", [
            {
                no: 2,
                name: "type",
                kind: "enum",
                T: () => [
                    "rustplus.AppTeamInfo.Note.Type",
                    AppTeamInfo_Note_Type
                ]
            },
            { no: 3, name: "x", kind: "scalar", T: 2 /*ScalarType.FLOAT*/ },
            { no: 4, name: "y", kind: "scalar", T: 2 /*ScalarType.FLOAT*/ },
            {
                no: 5,
                name: "icon",
                kind: "enum",
                T: () => [
                    "rustplus.AppTeamInfo.Note.Icon",
                    AppTeamInfo_Note_Icon
                ]
            },
            {
                no: 6,
                name: "colourIndex",
                kind: "enum",
                T: () => [
                    "rustplus.AppTeamInfo.Note.ColourIndex",
                    AppTeamInfo_Note_ColourIndex
                ]
            },
            {
                no: 7,
                name: "label",
                kind: "scalar",
                opt: true,
                T: 9 /*ScalarType.STRING*/
            }
        ]);
    }
    create(value?: PartialMessage<AppTeamInfo_Note>): AppTeamInfo_Note {
        const message = globalThis.Object.create(this.messagePrototype);
        message.type = 0;
        message.x = 0;
        message.y = 0;
        message.icon = 0;
        message.colourIndex = 0;
        if (value !== undefined)
            reflectionMergePartial<AppTeamInfo_Note>(this, message, value);
        return message;
    }
    internalBinaryRead(
        reader: IBinaryReader,
        length: number,
        options: BinaryReadOptions,
        target?: AppTeamInfo_Note
    ): AppTeamInfo_Note {
        const message = target ?? this.create(),
            end = reader.pos + length;
        while (reader.pos < end) {
            const [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* rustplus.AppTeamInfo.Note.Type type */ 2:
                    message.type = reader.int32();
                    break;
                case /* float x */ 3:
                    message.x = reader.float();
                    break;
                case /* float y */ 4:
                    message.y = reader.float();
                    break;
                case /* rustplus.AppTeamInfo.Note.Icon icon */ 5:
                    message.icon = reader.int32();
                    break;
                case /* rustplus.AppTeamInfo.Note.ColourIndex colourIndex */ 6:
                    message.colourIndex = reader.int32();
                    break;
                case /* optional string label */ 7:
                    message.label = reader.string();
                    break;
                default:
                    const u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(
                            `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`
                        );
                    const d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(
                            this.typeName,
                            message,
                            fieldNo,
                            wireType,
                            d
                        );
            }
        }
        return message;
    }
    internalBinaryWrite(
        message: AppTeamInfo_Note,
        writer: IBinaryWriter,
        options: BinaryWriteOptions
    ): IBinaryWriter {
        /* rustplus.AppTeamInfo.Note.Type type = 2; */
        if (message.type !== 0)
            writer.tag(2, WireType.Varint).int32(message.type);
        /* float x = 3; */
        if (message.x !== 0) writer.tag(3, WireType.Bit32).float(message.x);
        /* float y = 4; */
        if (message.y !== 0) writer.tag(4, WireType.Bit32).float(message.y);
        /* rustplus.AppTeamInfo.Note.Icon icon = 5; */
        if (message.icon !== 0)
            writer.tag(5, WireType.Varint).int32(message.icon);
        /* rustplus.AppTeamInfo.Note.ColourIndex colourIndex = 6; */
        if (message.colourIndex !== 0)
            writer.tag(6, WireType.Varint).int32(message.colourIndex);
        /* optional string label = 7; */
        if (message.label !== undefined)
            writer.tag(7, WireType.LengthDelimited).string(message.label);
        const u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(
                this.typeName,
                message,
                writer
            );
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message rustplus.AppTeamInfo.Note
 */
export const AppTeamInfo_Note = new AppTeamInfo_Note$Type();